Keep `appIdentity.ts` frozen so updates reach the phones
`app.config.ts` imports `@langx/shared/appIdentity` by path, so
`@expo/fingerprint` hashes that file's contents into the runtime version —
and EAS only hands an update to a binary whose runtime version matches.

#1279 added `verifyEmailUrl` there. Nothing in the config calls it, CI was
green and each publish succeeded, but the runtime version moved: the shipped
2.2 builds hold `8960664c…` and every update published since carries
`4fd7f8f8…`. Three merges' worth of updates went to a runtime version no
phone has. The symptom of this mistake is silence, which is what makes it
worth a rule.

So the function moves to `emailLinks.ts`, a sibling that may import
`appIdentity` freely, and `appIdentity.ts` goes back to the bytes the shipped
build was fingerprinted from — `@expo/fingerprint` confirms the hash returns
to its pre-#1279 value, so the update this merge publishes reaches 2.2 again.
`docs/decisions.md` carries the rule and the two commands that diagnose it.
