+++ packages/shared/src/index.ts
@@ -3,6 +3,7 @@ export * from './accountAge'
 export * from './age'
 export * from './appConfig'
 export * from './appIdentity'
+export * from './emailLinks'
 export * from './avatar'
 export * from './appScheme'
 export * from './badges'
